Set 0.0.0.0 in ALLOWED_HOSTS to avoid Django security error
Please read the instructions on how to deploy into production and how
to set ALLOWED_HOSTS carefully. Since other values like SECRET_KEY are
also straight up committed into the repository I assume preparing the
website for production and reading with some of the security
considerations of running a website nowadays will be done later.

For now this will allow the port forwarding to work and the host machine
to connect to the website using http://0.0.0.0:8000